Remove hardcoded base URL, add as an init option
